Fix display inactive selection
get_inactive_selection just return the text but doesn't fill the popdown with
the inactive selection. So the popdown must be re-filled with the selection and
the inactive selection, the inactive selection will be removed on the next
display with an other value.

issue4660
review15311003
(grafted from 5d126d0bd577af841b3a789ebe9e76a98e1a0a9b)