Fix TypeError during min/max computation in Graph
If the field is a datetime, a date or a timedelta but the value if None, it
could be compared to an existing instance and so raise a TypeError.
So the min/max value must be set with a correct value and TypeError must be
catched if the record value is None.

issue4888
review16501002
(grafted from 17e9572b187ced8ba91d9fa91280e8e2ff7da5a8)