Improve speed of setting default value on One2Many
When the default value of One2Many is a list of values, a lot of signals are
triggered for each record, this can be avoided like in changeset f09978373d97.
Also adding fields could always set_default to new record without signal and
just trigger the signal for the last record.
Finally, the validation in set_default is useless when performed from a
One2Many because the parent record will validate it anyway.

issue5161
review16131003