Fix setting twice the placeholder text
The implementation of placeholder for GTK+-2 must manage the case when the
placeholder is set multiple times. So when setting placeholder, if it does not
have the focus and it contains the default placeholder, the text should be
removed to have _focus_out set the new placeholder.

issue6137
review25921003
(grafted from b640f28da4d565e64c901cff11eb34a4fb8ce76d)