Skip validate only on constant readonly fields
A field can be readonly by the domain inversion in the validate method. So it
should be validated again in case its domain inversion change and reset it a
non-readonly.
So only field that are always readonly can be skipped for the validation.

issue6448
review37351002
(grafted from 70185ef5c14deda74932d6cabba19e283ef4c346)