Raise exception when the Lock for Login could not be acquired
The case when RPCExecute returns None is always forgotten. This shows that it
is the wrong design. Instead catching RPCException is always managed.
So it is better to raise an exception instead of returning None when the lock
for Login can not be acquired.

issue7211
review43231002
(grafted from 13fffff09a21b72953434eac50c01460fddee64e)