Remove decoration on NoModal and resize/center when shown
Since changeset 4655eba14d71, the size of the dialog is the same as the parent
but if the OS put decoration on the dialog, it position overflow the parent
which may hide the bottom where the button are.
The choice is to ensure request no decoration and to ensure that the size and
position are reset when the NoModal is shown.

issue7300
review47261002