Do not recycle first record as product result
The value of the field of the first record, which is edited in the list, will
be lost. Because the editable widget is started before the search pop-up and
it override the new default value with its content when the pop-up is closed.
So we can not reuse record create before the pop-up.

issue7941
review49681002
(grafted from 2ddba0a8db335de17cd09cb11b11c6654bad075a)