Rely on editing property of renderer to set value
In Python3, the 'remove-widget' is not trigger when the selected row is
changed. This makes the attribute editabletree_entry not reliable. Instead we
should use the editing property of the renderer and manage the editable
reference on the widget (which is a cleaner separation concept). This allows
also to remove duplicate code in set_value by calling on_editing_done.

issue7936
review46711002