Always display tree views to keep model up to date
Any tree view may be used even if it is not shown e.g. to save tree state. So
the model of TreeView which keeps a reference to the group must be updated if
the group has changed.

issue7885
review60491002
(grafted from 9e9ae0cba1c352bb95ef1268bad1abf904e17792)