Allow cells to have multiple renderers
Each cell can define prefixes and suffixes renderer which will be added to the
column using their expand property.
We use this new feature to replace the CellRendererBinary by
CellRendererClickablePixbuf's for each buttons.
We implement also an editable cell for Reference fields by using a Selection
and a Many2One cell.
In order to be able to navigate between cells of the same column, we do not
validate anymore the field when the cell is left for another of the same
record (this is a similar behavior to the form). Also the next and previous
columns are checked if they are editable if start_editing is set to ensure to
stay in edition mode during the all navigation.

issue8112
issue6918
review46761002