Use group _context to avoid leaking session context
When launching a new action, the context should not contain the session
context because it is automatically appended to the evaluation.
When creating a new group when the context model changed, it should not
contain the session context neither.

issue8312
review267301002
(grafted from fc20ef244ecadfcfe5393bd8ece163ab1a5d337e)