Do not rely on field views to load form fields
If the views attribute of the field is wrong or empty, the record will not
load the value for this field. The views attribute can be empty if it was
created by copying the field from its parent.
So it is safer to not rely on this attribute which is there only for
optimization and instead always use all the fields defined on the view.

issue8289
review263471002
(grafted from a2a957b7407f8d1b969842a20c6435818fee4afc)