Remove id from screen context and compare proper attributes
As mode is not always filled, we must compare only if view_ids is empty and we
must use the default value as fall-back.
The limit is a local parameter that should not be forced when opening an URL.
The screen context may be modified thanks to the context model, we must compare
the current context of the screen instead of the initial one.

issue8977
review272751003
(grafted from 70cdf0ef43d71517a6bd80fd3b3705d1fec4a536)