theme: Do not "readonly" editable comboboxes in tree views
We are using very "wide" flags for readonly (*.readonly),
so for a combobox that is inside an editable tree of a
readonly O2M field, the applied css is that of a readonly
combobox, even though it itself is not readonly.

So in this particular case, we restore the non-readonly
behaviour as much as possible.

Ideally we should make sure the "*.readonly" is properly
limited, but we rely on it being broad for a lot of cases
so...

Fix #16422